+++ src/cmd/send.rs
@@ -6,7 +6,7 @@ use crate::{
 use anyhow::Result;
 use clap::{Arg, ArgMatches};
 
-pub fn send_args() -> [Arg<'static, 'static>; 9] {
+pub fn send_args() -> [Arg<'static, 'static>; 10] {
     [
         Arg::with_name(arg::SENDER)
             .index(1)
@@ -37,6 +37,11 @@ pub fn send_args() -> [Arg<'static, 'static>; 9] {
             .takes_value(true)
             .required_unless_one(&[arg::SUBJECT, arg::CONTENT])
             .help("Path of the message file"),
+        Arg::with_name(arg::ARCHIVE_DIR)
+            .long(arg::ARCHIVE_DIR)
+            .takes_value(true)
+            .default_value("./archived_messages")
+            .help("Path of archived messages"),
         Arg::with_name(arg::DISPLAY)
             .long(arg::DISPLAY)
             .takes_value(false)